venta-de-garage-web/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
email-validator==1.1.3
pydantic[email]==1.8.2
aiosqlite==0.19.0
Pillow==9.5.0
requests==2.31.0
aiofiles
werkzeug
